Improve wording and structure of paragraph.
To be grammatically correct, the phrase "Confusion about what meaning a T* is the source" should have had a doubled "is", which would have been awkward, so the sentence has been reworded.

The second and third sentences were inverted. Prior to the inversion, the sentence starting with "For example" initially appeared to be expanding on the debuggability of not_null, rather than providing an example of how using separate names improves code quality.