Drop odd editorial note.
Many of the guidelines in here are contradictory to legacy behavior,
they don't get similar notes about "billions of lines that contradict."

"This code cannot all be rewritten ..." - I actually think this is
false, we've been doing a lot of large-scale refactoring, and this is
pretty solidly within the realm of what I think we *could* do.

The example of owning raw pointers in resource handles (specifically
unique_ptr and shared_ptr) is pretty minor, and pretty easily handled by
just not rewriting code in the standard library itself.

The ABI issue seems (as discussed elsewhere) potentially to be out of
scope or misleading.

To my mind, seems best to drop the whole note.