ES.64: Don't use the comma operator
In C++, (a,b) evaluates to b. This is the "comma operator".

Most of the time the comma operator makes the code unreadable; in the worst case the comma operator is just an unintended typo in the code. There are only very few legitimate cases, such as the initialization or increase expression of for statements, where the comma operator really increases the readability of the code.

See the suggested rule.