F.21 and GSL and Etc.
The cleanup continues.
Note the caller-allocates example in F.21.
Note the attempted cleanup of the Enforcement section of F.7: too many
ands and ors; did I get it right?
Note the simplification of the treatment of char* in the GSL section;
char* is considered a pointer to a single char, rather than a C-style
string; viable?