In.force: wording
An unenforceable rule isn't, by definition, a rule. I believe it meant to be rules without mechanical enforcement.

The first part of the modified disjunction doesn't make sense. It has the problem mentioned above, and also mentions that "all is only a part". Thus, I divided the disjunction and applied the same fix as above to the first part.