Closed 856 and 857
Added C++17-specific text in ES.6 for `if` and `switch` initializers.

Note that the Guidelines assume C++14, so the "Enforcement" clauses for
C++17-specific rules will be separate until we assume people have broad
access to C++17 features in their compilers. Once that happens we can
roll the C++17-specific parts into the main text.