Fix CP.3 to allow local reasoning and compilation without error
(CP.3: Minimize explicit sharing of writable data)

Use 'ES.28: Use lambdas for complex initialization,
especially of const variables' to ensure we can see from the body
of the function that the function called in async only require const
parameter.

This ensure that if a non local change to the function signature
(making the const parameter non const) it will result in a compilation
error.

Added necessary construct and compiled with https://godbolt.org/g/tjGXbV