Rename rnd to round in I.1
At first, I assumed that this is "rand" function, not "round" function,
and was confused why exactly is it a problem that this returns
different values. While it's easy to notice when actually reading the
code, this slows down the comprehension, and bad function naming is
not point of that particular guideline.