Explain freedoms available in F.16-F.18
As I discussed in #1407, I found the existing wording confusing as I
expected that the guidelines would provide me with one recommended way
to author function arguments in all situations.  However, the
discussion on the issue clarified that the intent was to allow many
ways to author functions, and only to disallow specific clearly
non-optimal cases.

This is my attempt to reduce that confusion, by explicitly calling out
when such freedoms exist.