[SF.10] Simplify example
The original example uses a std header to demonstrate transitive
inclusion, which makes it relate to real code, on the one hand. Yet, the
header is too long to be put in the example, so it's omitted and thus
the example is rather obscure, on the other hand.

Switch to a constructed example which uses a short header which is
presented right there to make the point clearer.