F.60: Remove C-style cast (T&) from example of invalid C++
The C-style cast in the example of constructing a `nullptr` reference appears unnecessary, in "F.60: Prefer T* over T& when “no argument” is a valid option". The example would also compile without any cast.

Of course, `T* p = nullptr; T& r = *p;` is still not valid C++, but that is according to the intention of the example.