Modify R.34 and delete R.36 to conform to F.16-F.18
As it stands now, to my understanding R.34 conflicts with F.16-F.18.

R.34 suggests taking `shared_ptr` by value in a "sink" function,
but F.18 mentions explicitly that these sort of functions should pass
by `&&`, and specifically _not_ by value.  This is mentioned by @hsutter in
[this comment](https://github.com/isocpp/CppCoreGuidelines/issues/1916#issuecomment-1183468209)

So, this PR attempts to resolve the conflict by preferring the
"F.16-F.18" guidelines and attempting to modify the R guidelines to fit.

There are a number of other ways to avoid this ambiguity, for example
we could explicitly state that `shared_ptr` is not subject to
F.16-F.18 somewhere in those guidelines.  R.34 could also be made to
fit if it were explicitly stated that `shared_ptr` be considered
"cheap to copy", although I don't see how R.36 could be made to fit
with F.16-F.18.

I don't have any strong opinion on what the "right" resolution is
here, but the fact that these two sets of guidelines seem to disagree
has been a source of confusion.