Add guideline to not invoke undefined behavior
At my CppCon 2022 talk [
Purging Undefined Behavior & Intel Assumptions in a Legacy C++
Codebase](https://www.youtube.com/watch?v=vEtGtphI3lc) I spoke about a
guideline we added to our fork of the Core Guidelines to not invoke
undefined behavior.  I got a question after the talk: "why is this not
part of the core guidelines?"

Now that I am giving a new version of this talk at Meeting C++ I was
inspired by this question to propose this as an upstream guideline.

The current wording is our internal wording that might need some
upgrades but first I wanted to know if there was interest in accepting
this guideline before doing further refinement of the text.