[P.6] copyedit: Move a comment for consistency, change "a" to "the", and fix a typo: "f" to "f2" (#2288)
* [P.6] trivial: s/f/f2/ (#2288)

A comment in a code example was intended to refer to a function named "f2",
but the comment uses the name "f" instead, presumably because the code was
copied from the previous example; this commit fixes the typo.

* [P.6] trivial: Move a code comment (//...) to the line above the code (#2288)

This matches the way the previous, related code example is written,
a fact you can observe more readily by reviewing the patch of this
commit with larger context:

    git diff -U19

Specifically, the previous example looks like this:

    // bad: the number of elements is not passed to f()
    f(new int[n]);

* [P.6] trivial: s/a/the/ (#2288)

While "a wrong number can be passed" makes sense (and there are perhaps
good philosophical reasons to prefer it), I believe that the more natural
way to say it in common English is "the wrong number can be passed"; thus,
this commit changes "a" to "the" in the comment of a code example.

---------

Co-authored-by: Herb Sutter <herb.sutter@gmail.com>